Add files via upload
Protein inference, the identification of the protein set that is the origin of
a given peptide profile, is a fundamental challenge in proteomics. In this
work, we present a solution based on machine learning that predicts the
protein set from a proteomics mixture, given the sequence universe of
possible proteins and a target peptide profile. The results are tested on
both simulated datasets and real datasets.